requirements.txt
absl-py==0.7.1
aniso8601==8.0.0
astor==0.8.0
atomicwrites==1.2.1
attrs==18.2.0
awscli==1.16.232
boto==2.49.0
boto3==1.9.228
botocore==1.12.228
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
Django==2.1.4
docutils==0.15.2
Faker==1.0.7
Flask==1.1.1
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
gast==0.2.2
glob3==0.0.1
grpcio==1.21.1
h5py==2.9.0
idna==2.8
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mock==3.0.5
more-itertools==4.3.0
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.24.0
Pillow==5.3.0
pluggy==0.8.0
protobuf==3.8.0
py==1.7.0
pyasn1==0.4.7
pycodestyle==2.5.0
pydocstyle==4.0.1
pymongo==3.9.0
pyparsing==2.3.0
pyrsistent==0.15.4
pytest==3.9.2
python-dateutil==2.7.5
pytz==2018.7
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
scipy==1.3.0
six==1.11.0
snowballstemmer==1.9.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
text-unidecode==1.2
urllib3==1.25.3
virtualenv==16.0.0
Werkzeug==0.15.4
WTForms==2.2.1
